ffmpeg_sys_next

枚举 AVFrameSideDataType

#[repr(i32)]
pub enum AVFrameSideDataType {

    AV_FRAME_DATA_PANSCAN = 0,
    AV_FRAME_DATA_A53_CC = 1,
    AV_FRAME_DATA_STEREO3D = 2,
    AV_FRAME_DATA_MATRIXENCODING = 3,
    AV_FRAME_DATA_DOWNMIX_INFO = 4,
    AV_FRAME_DATA_REPLAYGAIN = 5,
    AV_FRAME_DATA_DISPLAYMATRIX = 6,
    AV_FRAME_DATA_AFD = 7,
    AV_FRAME_DATA_MOTION_VECTORS = 8,
    AV_FRAME_DATA_SKIP_SAMPLES = 9,
    AV_FRAME_DATA_AUDIO_SERVICE_TYPE = 10,
    AV_FRAME_DATA_MASTERING_DISPLAY_METADATA = 11,
    AV_FRAME_DATA_GOP_TIMECODE = 12,
    AV_FRAME_DATA_SPHERICAL = 13,
    AV_FRAME_DATA_CONTENT_LIGHT_LEVEL = 14,
    AV_FRAME_DATA_ICC_PROFILE = 15,
    AV_FRAME_DATA_S12M_TIMECODE = 16,
    AV_FRAME_DATA_DYNAMIC_HDR_PLUS = 17,
    AV_FRAME_DATA_REGIONS_OF_INTEREST = 18,
    AV_FRAME_DATA_VIDEO_ENC_PARAMS = 19,
    AV_FRAME_DATA_SEI_UNREGISTERED = 20,
    AV_FRAME_DATA_FILM_GRAIN_PARAMS = 21,
    AV_FRAME_DATA_DETECTION_BBOXES = 22,
    AV_FRAME_DATA_DOVI_RPU_BUFFER = 23,
    AV_FRAME_DATA_DOVI_METADATA = 24,
    AV_FRAME_DATA_DYNAMIC_HDR_VIVID = 25,
    AV_FRAME_DATA_AMBIENT_VIEWING_ENVIRONMENT = 26,
    AV_FRAME_DATA_VIDEO_HINT = 27,
    AV_FRAME_DATA_LCEVC = 28,
    AV_FRAME_DATA_VIEW_ID = 29,
    AV_FRAME_DATA_3D_REFERENCE_DISPLAYS = 30,
    AV_FRAME_DATA_EXIF = 31,

}

@defgroup lavu_frame AVFrame @ingroup lavu_data

@{ AVFrame is an abstraction for reference-counted raw multimedia data.

Variants

AV_FRAME_DATA_PANSCAN = 0

The data is the AVPanScan struct defined in libavcodec.

AV_FRAME_DATA_A53_CC = 1

ATSC A53 Part 4 Closed Captions. A53 CC bitstream is stored as uint8_t in AVFrameSideData.data. The number of bytes of CC data is AVFrameSideData.size.

AV_FRAME_DATA_STEREO3D = 2

Stereoscopic 3d metadata. The data is the AVStereo3D struct defined in libavutil/stereo3d.h.

AV_FRAME_DATA_MATRIXENCODING = 3

The data is the AVMatrixEncoding enum defined in libavutil/channel_layout.h.

AV_FRAME_DATA_DOWNMIX_INFO = 4

Metadata relevant to a downmix procedure. The data is the AVDownmixInfo struct defined in libavutil/downmix_info.h.

AV_FRAME_DATA_REPLAYGAIN = 5

ReplayGain information in the form of the AVReplayGain struct.

AV_FRAME_DATA_DISPLAYMATRIX = 6

This side data contains a 3x3 transformation matrix describing an affine transformation that needs to be applied to the frame for correct presentation.

See libavutil/display.h for a detailed description of the data.

AV_FRAME_DATA_AFD = 7

Active Format Description data consisting of a single byte as specified in ETSI TS 101 154 using AVActiveFormatDescription enum.

AV_FRAME_DATA_MOTION_VECTORS = 8

Motion vectors exported by some codecs (on demand through the export_mvs flag set in the libavcodec AVCodecContext flags2 option). The data is the AVMotionVector struct defined in libavutil/motion_vector.h.

AV_FRAME_DATA_SKIP_SAMPLES = 9

Recommends skipping the specified number of samples. This is exported only if the “skip_manual” AVOption is set in libavcodec. This has the same format as AV_PKT_DATA_SKIP_SAMPLES. @code u32le number of samples to skip from start of this packet u32le number of samples to skip from end of this packet u8 reason for start skip u8 reason for end skip (0=padding silence, 1=convergence) @endcode

AV_FRAME_DATA_AUDIO_SERVICE_TYPE = 10

This side data must be associated with an audio frame and corresponds to enum AVAudioServiceType defined in avcodec.h.

AV_FRAME_DATA_MASTERING_DISPLAY_METADATA = 11

Mastering display metadata associated with a video frame. The payload is an AVMasteringDisplayMetadata type and contains information about the mastering display color volume.

AV_FRAME_DATA_GOP_TIMECODE = 12

The GOP timecode in 25 bit timecode format. Data format is 64-bit integer. This is set on the first frame of a GOP that has a temporal reference of 0.

AV_FRAME_DATA_SPHERICAL = 13

The data represents the AVSphericalMapping structure defined in libavutil/spherical.h.

AV_FRAME_DATA_CONTENT_LIGHT_LEVEL = 14

Content light level (based on CTA-861.3). This payload contains data in the form of the AVContentLightMetadata struct.

AV_FRAME_DATA_ICC_PROFILE = 15

The data contains an ICC profile as an opaque octet buffer following the format described by ISO 15076-1 with an optional name defined in the metadata key entry “name”.

AV_FRAME_DATA_S12M_TIMECODE = 16

Timecode which conforms to SMPTE ST 12-1. The data is an array of 4 uint32_t where the first uint32_t describes how many (1-3) of the other timecodes are used. The timecode format is described in the documentation of av_timecode_get_smpte_from_framenum() function in libavutil/timecode.h.

AV_FRAME_DATA_DYNAMIC_HDR_PLUS = 17

HDR dynamic metadata associated with a video frame. The payload is an AVDynamicHDRPlus type and contains information for color volume transform - application 4 of SMPTE 2094-40:2016 standard.

AV_FRAME_DATA_REGIONS_OF_INTEREST = 18

Regions Of Interest, the data is an array of AVRegionOfInterest type, the number of array element is implied by AVFrameSideData.size / AVRegionOfInterest.self_size.

AV_FRAME_DATA_VIDEO_ENC_PARAMS = 19

Encoding parameters for a video frame, as described by AVVideoEncParams.

AV_FRAME_DATA_SEI_UNREGISTERED = 20

User data unregistered metadata associated with a video frame. This is the H.26[45] UDU SEI message, and shouldn’t be used for any other purpose The data is stored as uint8_t in AVFrameSideData.data which is 16 bytes of uuid_iso_iec_11578 followed by AVFrameSideData.size - 16 bytes of user_data_payload_byte.

AV_FRAME_DATA_FILM_GRAIN_PARAMS = 21

Film grain parameters for a frame, described by AVFilmGrainParams. Must be present for every frame which should have film grain applied.

May be present multiple times, for example when there are multiple alternative parameter sets for different video signal characteristics. The user should select the most appropriate set for the application.

AV_FRAME_DATA_DETECTION_BBOXES = 22

Bounding boxes for object detection and classification, as described by AVDetectionBBoxHeader.

AV_FRAME_DATA_DOVI_RPU_BUFFER = 23

Dolby Vision RPU raw data, suitable for passing to x265 or other libraries. Array of uint8_t, with NAL emulation bytes intact.

AV_FRAME_DATA_DOVI_METADATA = 24

Parsed Dolby Vision metadata, suitable for passing to a software implementation. The payload is the AVDOVIMetadata struct defined in libavutil/dovi_meta.h.

AV_FRAME_DATA_DYNAMIC_HDR_VIVID = 25

HDR Vivid dynamic metadata associated with a video frame. The payload is an AVDynamicHDRVivid type and contains information for color volume transform - CUVA 005.1-2021.

AV_FRAME_DATA_AMBIENT_VIEWING_ENVIRONMENT = 26

Ambient viewing environment metadata, as defined by H.274.

AV_FRAME_DATA_VIDEO_HINT = 27

Provide encoder-specific hinting information about changed/unchanged portions of a frame. It can be used to pass information about which macroblocks can be skipped because they didn’t change from the corresponding ones in the previous frame. This could be useful for applications which know this information in advance to speed up encoding.

AV_FRAME_DATA_LCEVC = 28

Raw LCEVC payload data, as a uint8_t array, with NAL emulation bytes intact.

AV_FRAME_DATA_VIEW_ID = 29

This side data must be associated with a video frame. The presence of this side data indicates that the video stream is composed of multiple views (e.g. stereoscopic 3D content, cf. H.264 Annex H or H.265 Annex G). The data is an int storing the view ID.

AV_FRAME_DATA_3D_REFERENCE_DISPLAYS = 30

This side data contains information about the reference display width(s) and reference viewing distance(s) as well as information about the corresponding reference stereo pair(s), i.e., the pair(s) of views to be displayed for the viewer’s left and right eyes on the reference display at the reference viewing distance. The payload is the AV3DReferenceDisplaysInfo struct defined in libavutil/tdrdi.h.

AV_FRAME_DATA_EXIF = 31

Extensible image file format metadata. The payload is a buffer containing EXIF metadata, starting with either 49 49 2a 00, or 4d 4d 00 2a.