ffmpeg_sys_next

函数 av_find_best_pix_fmt_of_2

pub unsafe extern "C" fn av_find_best_pix_fmt_of_2(
    dst_pix_fmt1: AVPixelFormat,
    dst_pix_fmt2: AVPixelFormat,
    src_pix_fmt: AVPixelFormat,
    has_alpha: c_int,
    loss_ptr: *mut c_int,
) -> AVPixelFormat

Compute what kind of losses will occur when converting from one specific pixel format to another. When converting from one pixel format to another, information loss may occur. For example, when converting from RGB24 to GRAY, the color information will be lost. Similarly, other losses occur when converting from some formats to other formats. These losses can involve loss of chroma, but also loss of resolution, loss of color depth, loss due to the color space conversion, loss of the alpha bits or loss due to color quantization. av_get_fix_fmt_loss() informs you about the various types of losses which will occur when converting from one pixel format to another.

@param[in] dst_pix_fmt destination pixel format @param[in] src_pix_fmt source pixel format @param[in] has_alpha Whether the source pixel format alpha channel is used. @return Combination of flags informing you what kind of losses will occur (maximum loss for an invalid dst_pix_fmt).