ffmpeg_sys_next

函数 av_frame_is_writable

pub unsafe extern "C" fn av_frame_is_writable(
    frame: *mut AVFrame,
) -> c_int

Check if the frame data is writable.

@return A positive value if the frame data is writable (which is true if and only if each of the underlying buffers has only one reference, namely the one stored in this frame). Return 0 otherwise.

If 1 is returned the answer is valid until av_buffer_ref() is called on any of the underlying AVBufferRefs (e.g. through av_frame_ref() or directly).

@see av_frame_make_writable(), av_buffer_is_writable()