ffmpeg_sys_next

函数 av_frame_new_side_data

pub unsafe extern "C" fn av_frame_new_side_data(
    frame: *mut AVFrame,
    type_: AVFrameSideDataType,
    size: usize,
) -> *mut AVFrameSideData

Add a new side data to a frame.

@param frame a frame to which the side data should be added @param type type of the added side data @param size size of the side data

@return newly added side data on success, NULL on error