ffmpeg_sys_next

函数 av_opt_get

pub unsafe extern "C" fn av_opt_get(
    obj: *mut c_void,
    name: *const c_char,
    search_flags: c_int,
    out_val: *mut *mut u8,
) -> c_int

@defgroup opt_get_funcs Option getting functions @{ Those functions get a value of the option with the given name from an object.

@param[in] obj a struct whose first element is a pointer to an AVClass. @param[in] name name of the option to get. @param[in] search_flags flags passed to av_opt_find2. I.e. if AV_OPT_SEARCH_CHILDREN is passed here, then the option may be found in a child of obj. @param[out] out_val value of the option will be written here @return >=0 on success, a negative error code otherwise / /** @note the returned string will be av_malloc()ed and must be av_free()ed by the caller

@note if AV_OPT_ALLOW_NULL is set in search_flags in av_opt_get, and the option is of type AV_OPT_TYPE_STRING, AV_OPT_TYPE_BINARY or AV_OPT_TYPE_DICT and is set to NULL, *out_val will be set to NULL instead of an allocated empty string.