ffmpeg_sys_next

结构体 AVFilterLink

#[repr(C)]
pub struct AVFilterLink {

    pub src: *mut AVFilterContext,
    pub srcpad: *mut AVFilterPad,
    pub dst: *mut AVFilterContext,
    pub dstpad: *mut AVFilterPad,
    pub type_: AVMediaType,
    pub format: c_int,
    pub w: c_int,
    pub h: c_int,
    pub sample_aspect_ratio: AVRational,
    pub colorspace: AVColorSpace,
    pub color_range: AVColorRange,
    pub sample_rate: c_int,
    pub ch_layout: AVChannelLayout,
    pub time_base: AVRational,
    pub side_data: *mut *mut AVFrameSideData,
    pub nb_side_data: c_int,
    pub alpha_mode: AVAlphaMode,
    pub incfg: AVFilterFormatsConfig,
    pub outcfg: AVFilterFormatsConfig,

}

A link between two filters. This contains pointers to the source and destination filters between which this link exists, and the indexes of the pads involved. In addition, this link also contains the parameters which have been negotiated and agreed upon between the filter, such as image dimensions, format, etc.

Applications must not normally access the link structure directly. Use the buffersrc and buffersink API instead. In the future, access to the header may be reserved for filters implementation.

Fields
src: *mut AVFilterContext

< source filter

srcpad: *mut AVFilterPad

< output pad on the source filter

dst: *mut AVFilterContext

< dest filter

dstpad: *mut AVFilterPad

< input pad on the dest filter

type_: AVMediaType

< filter media type

format: c_int

< agreed upon media format

w: c_int

< agreed upon image width

h: c_int

< agreed upon image height

sample_aspect_ratio: AVRational

< agreed upon sample aspect ratio

colorspace: AVColorSpace

< agreed upon YUV color space

color_range: AVColorRange

< agreed upon YUV color range

sample_rate: c_int

< samples per second

ch_layout: AVChannelLayout

< channel layout of current buffer (see libavutil/channel_layout.h)

time_base: AVRational

Define the time base used by the PTS of the frames/samples which will pass through this link. During the configuration stage, each filter is supposed to change only the output timebase, while the timebase of the input link is assumed to be an unchangeable property.

side_data: *mut *mut AVFrameSideDatanb_side_data: c_intalpha_mode: AVAlphaMode

< alpha mode (for videos with an alpha channel)

incfg: AVFilterFormatsConfig

Lists of supported formats / etc. supported by the input filter.

outcfg: AVFilterFormatsConfig

Lists of supported formats / etc. supported by the output filter.