ffmpeg_sys_next

结构体 AVProgram

#[repr(C)]
pub struct AVProgram {

    pub id: c_int,
    pub flags: c_int,
    pub discard: AVDiscard,
    pub stream_index: *mut c_uint,
    pub nb_stream_indexes: c_uint,
    pub metadata: *mut AVDictionary,
    pub program_num: c_int,
    pub pmt_pid: c_int,
    pub pcr_pid: c_int,
    pub pmt_version: c_int,
    pub start_time: i64,
    pub end_time: i64,
    pub pts_wrap_reference: i64,
    pub pts_wrap_behavior: c_int,

}

New fields can be added to the end with minor version bumps. Removal, reordering and changes to existing fields require a major version bump. sizeof(AVProgram) must not be used outside libav*.

Fields
id: c_intflags: c_intdiscard: AVDiscard

< selects which program to discard and which to feed to the caller

stream_index: *mut c_uintnb_stream_indexes: c_uintmetadata: *mut AVDictionaryprogram_num: c_intpmt_pid: c_intpcr_pid: c_intpmt_version: c_intstart_time: i64

All fields below this line are not part of the public API. They may not be used outside of libavformat and can be changed and removed at will. New public fields should be added right above.

end_time: i64pts_wrap_reference: i64

< reference dts for wrap detection

pts_wrap_behavior: c_int

< behavior on wrap detection