ffmpeg_sys_next

结构体 AVStreamGroupTileGrid__bindgen_ty_1

#[repr(C)]
pub struct AVStreamGroupTileGrid__bindgen_ty_1 {
    pub idx: c_uint,
    pub horizontal: c_int,
    pub vertical: c_int,
}

An @ref nb_tiles sized array of offsets in pixels from the topleft edge of the canvas, indicating where each stream should be placed. It must be allocated with the av_malloc() family of functions.

demuxing: set by libavformat, must not be modified by the caller.

muxing: set by the caller before avformat_write_header().

Freed by libavformat in avformat_free_context().

Fields
idx: c_uint

Index of the stream in the group this tile references.

Must be < @ref AVStreamGroup.nb_streams “nb_streams”.

horizontal: c_int

Offset in pixels from the left edge of the canvas where the tile should be placed.

vertical: c_int

Offset in pixels from the top edge of the canvas where the tile should be placed.